refactor: extract calculateMaxLevelsPerSide function to perpLayoutUtils for better code organization
- Moved the calculateMaxLevelsPerSide function from PerpDesktopLayout components to a new utility file, perpLayoutUtils, to enhance code modularity and maintainability.
- Updated references in PerpDesktopLayout and PerpDesktopLayout.web to use the new utility function.
- Adjusted orderBook width and trading dimensions in PERP_LAYOUT_CONFIG for improved layout consistency.
