+++ packages/kit/src/views/Perp/layouts/PerpDesktopLayout.web.tsx
@@ -25,37 +25,7 @@ import {
   PerpAccountPanel,
 } from '../components/TradingPanel/panels/PerpAccountPanel';
 import { PerpTradingPanel } from '../components/TradingPanel/PerpTradingPanel';
-
-function calculateMaxLevelsPerSide(containerHeight: number): number {
-  // We want to minimize visible "blank" space at the bottom of the order book.
-  // The vertical web order book renders:
-  // - Root padding: 1px top + 1px bottom (2px)
-  // - Table header (Price/Size/Total): 24px
-  // - Spread row: 24px + 1px marginTop (25px)
-  // - Each level row: 24px + 1px marginTop (25px)
-  //
-  // Total height: baseHeight(51px) + 2 * levelsPerSide * 25px
-  const baseHeight = 2 + 24 + 25;
-  const levelRowStep = 25;
-
-  if (containerHeight <= 0) return 11;
-  if (containerHeight <= baseHeight) return 3;
-
-  let levelsPerSide = Math.floor(
-    (containerHeight - baseHeight) / (2 * levelRowStep),
-  );
-  levelsPerSide = Math.max(3, Math.min(levelsPerSide, 50));
-
-  // If we have a noticeable gap, prefer one extra level and accept a tiny clip
-  // instead of showing empty space.
-  const usedHeight = baseHeight + 2 * levelsPerSide * levelRowStep;
-  const blank = containerHeight - usedHeight;
-  if (blank > levelRowStep / 2 && levelsPerSide < 50) {
-    levelsPerSide += 1;
-  }
-
-  return levelsPerSide;
-}
+import { calculateMaxLevelsPerSide } from './perpLayoutUtils';
 
 function PerpDesktopLayout() {
   const intl = useIntl();
@@ -64,7 +34,7 @@ function PerpDesktopLayout() {
 
   const layout = PERP_LAYOUT_CONFIG.desktop;
 
-  const showOrderBook = gtXl && layoutState.orderBook.visible;
+  const showOrderBook = gtXl && (layoutState.orderBook?.visible ?? true);
   const tradingWidth = gtXl ? layout.widths.tradingXl : layout.widths.trading;
   const orderBookMaxLevelsPerSide = useMemo(
     () =>
